fix: onboarding flow
