+++ lib/routes/routes_name.dart
@@ -8,4 +8,5 @@ abstract class RoutesName {
   static const onboarding = '/onboarding';
   static const introduction = '/introduction';
   static const dream = '/dream';
+  static const onboardingFlow = '/onboarding-flow';
 }
