--- lib/view/dream_page.dart
@@ -1,172 +1,0 @@
-// lib/view/dream_screen.dart
-
-import 'package:flutter/material.dart';
-
-class DreamScreen extends StatefulWidget {
-  const DreamScreen({super.key});
-
-  @override
-  State<DreamScreen> createState() => _DreamScreenState();
-}
-
-class _DreamScreenState extends State<DreamScreen> {
-  final TextEditingController _universityController = TextEditingController();
-  final TextEditingController _majorController = TextEditingController();
-
-  @override
-  void dispose() {
-    _universityController.dispose();
-    _majorController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    const Color primaryBlue = Color(0xFF183D8D);
-
-    return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        backgroundColor: Colors.transparent,
-        elevation: 0,
-        leading: IconButton(
-          icon: const Icon(Icons.arrow_back_ios, color: Colors.black),
-          onPressed: () => Navigator.pop(context),
-        ),
-        title: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            // Previous screens (empty dots)
-            Container(
-              width: 8.0,
-              height: 8.0,
-              decoration: BoxDecoration(
-                color: Colors.grey.shade300,
-                shape: BoxShape.circle,
-              ),
-            ),
-            const SizedBox(width: 8.0),
-            Container(
-              width: 8.0,
-              height: 8.0,
-              decoration: BoxDecoration(
-                color: Colors.grey.shade300,
-                shape: BoxShape.circle,
-              ),
-            ),
-            const SizedBox(width: 8.0),
-            // Current screen (filled dot)
-            Container(
-              width: 8.0,
-              height: 8.0,
-              decoration: const BoxDecoration(
-                color: primaryBlue,
-                shape: BoxShape.circle,
-              ),
-            ),
-          ],
-        ),
-        centerTitle: true,
-      ),
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 24.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: [
-              const SizedBox(height: 20), // Spacing from app bar
-
-              // Logo
-              Center(
-                child: Image.asset(
-                  'assets/images/logo.png',
-                  height: 60.0, // Smaller logo
-                ),
-              ),
-
-              const SizedBox(height: 30.0),
-
-              // Title
-              const Text(
-                'Ceritakan impianmu!',
-                textAlign: TextAlign.center,
-                style: TextStyle(
-                  fontSize: 24.0,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.black,
-                ),
-              ),
-
-              const SizedBox(height: 40.0),
-
-              // Dream University Input Field
-              _buildTextField(
-                controller: _universityController,
-                hintText: 'Universitas Impian',
-              ),
-
-              const SizedBox(height: 20.0),
-
-              // Dream Major Input Field
-              _buildTextField(
-                controller: _majorController,
-                hintText: 'Jurusan Impian',
-              ),
-
-              const Spacer(),
-
-              // "Ayo Mulai" Button
-              ElevatedButton(
-                onPressed: () {
-                  // TODO: Save dream data
-                  print('Dream University: ${_universityController.text}');
-                  print('Dream Major: ${_majorController.text}');
-                  // TODO: Navigate to the main app dashboard
-                  print('Ayo Mulai button pressed!');
-                },
-                style: ElevatedButton.styleFrom(
-                  backgroundColor: primaryBlue,
-                  foregroundColor: Colors.white,
-                  padding: const EdgeInsets.symmetric(vertical: 16.0),
-                  shape: const StadiumBorder(),
-                  textStyle: const TextStyle(
-                    fontSize: 18,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-                child: const Text('Ayo Mulai'),
-              ),
-
-              const SizedBox(height: 40.0), // Bottom padding
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  // Helper method to build consistent text fields
-  Widget _buildTextField({
-    required TextEditingController controller,
-    required String hintText,
-    TextInputType keyboardType = TextInputType.text,
-  }) {
-    return Container(
-      decoration: BoxDecoration(
-        color: Colors.grey[200],
-        borderRadius: BorderRadius.circular(10.0),
-      ),
-      child: TextField(
-        controller: controller,
-        keyboardType: keyboardType,
-        decoration: InputDecoration(
-          hintText: hintText,
-          hintStyle: const TextStyle(color: Colors.grey),
-          border: InputBorder.none,
-          contentPadding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 14.0),
-        ),
-        style: const TextStyle(fontSize: 16, color: Colors.black87),
-      ),
-    );
-  }
-}
